cap nhap layout admin, fix bug

+++ app/Http/Controllers/CategoryController.php
@@ -1,14 +1,19 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\Category;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index()
     {
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         $categorys = Category::all(); // Lấy tất cả sản phẩm từ cơ sở dữ liệu
 
         // Đây là số trang hiện tại, nếu bạn cần phân trang
@@ -22,10 +27,16 @@ class CategoryController extends Controller
     }
     public function create()
     {
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         return view('layout_add_category');
     }
     public function store(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         $request->validate([
             'TenDanhMuc' => 'required|string|max:255|unique:categories,TenDanhMuc',
         ]);
@@ -38,27 +49,36 @@ class CategoryController extends Controller
     }
     public function destroy($id)
     {
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         $category = Category::findOrFail($id);
         $category->delete();
 
         return redirect()->back()->with('success', 'Danh mục đã được xóa thành công!');
     }
     public function edit($id)
-{
-    $category = Category::findOrFail($id);
-    return view('layout_edit_category', compact('category'));
-}
+    {
+        if (!Auth::check()) {
+            return redirect('login');
+        }
+        $category = Category::findOrFail($id);
+        return view('layout_edit_category', compact('category'));
+    }
 
-public function update(Request $request, $id)
-{
-    $request->validate([
-        'TenDanhMuc' => 'required|string|max:255',
-    ]);
+    public function update(Request $request, $id)
+    {
+        if (!Auth::check()) {
+            return redirect('login');
+        }
+        $request->validate([
+            'TenDanhMuc' => 'required|string|max:255',
+        ]);
 
-    $category = Category::findOrFail($id);
-    $category->TenDanhMuc = $request->input('TenDanhMuc');
-    $category->save();
+        $category = Category::findOrFail($id);
+        $category->TenDanhMuc = $request->input('TenDanhMuc');
+        $category->save();
 
-    return redirect()->back();
-}
+        return redirect()->back();
+    }
 }
